added features needed for muxlink relocking

+++ src/SystemDependenceGraph/Graph.py
@@ -1,6 +1,7 @@
 from click import style
 import pygraphviz as pgv
 from SystemDependenceGraph.DependenceNode import *
+from os.path import exists
 
 class Graph:
 
@@ -37,16 +38,23 @@ class Graph:
     def get_nodes(self):
         return self.nodes
 
-    def print_graph(self, path):
+    def print_graph(self, path, index=0):
         G = pgv.AGraph(directed=True, strict=False)
         nlist = [str(n)+" mem: "+str(id(n)) for n in self.nodes]
 
         G.add_nodes_from(nlist)
-        feature_file = open(path+"_features.txt", 'w')
-        cell_file = open(path+"_cells.txt", 'w')
-        link_train_file = open(path+"_link_train.txt", 'w')
-        link_test_file = open(path+"_link_test.txt", 'w')
-        link_test_n_file = open(path+"_link_test_n.txt", 'w')
+        if exists(path+"_features.txt"):
+            feature_file = open(path+"_features.txt", 'a')
+            cell_file = open(path+"_cells.txt", 'a')
+            link_train_file = open(path+"_link_train.txt", 'a')
+            link_test_file = open(path+"_link_test.txt", 'a')
+            link_test_n_file = open(path+"_link_test_n.txt", 'a')
+        else:
+            feature_file = open(path+"_features.txt", 'w')
+            cell_file = open(path+"_cells.txt", 'w')
+            link_train_file = open(path+"_link_train.txt", 'w')
+            link_test_file = open(path+"_link_test.txt", 'w')
+            link_test_n_file = open(path+"_link_test_n.txt", 'w')
 
         for i, n in enumerate(self.nodes):
             print(n)
@@ -54,16 +62,16 @@ class Graph:
             print(n, file=cell_file)
             if isinstance(n, CondNode) and "locking_key" in n.cond_statement.get_cond_dependencies():
                 for c in n.true_statements:
-                    print(i, self.nodes.index(c), file=link_test_file)
+                    print(i+index, self.nodes.index(c)+index, file=link_test_file)
                 for c in n.false_statements:
-                    print(i, self.nodes.index(c), file=link_test_n_file)
+                    print(i+index, self.nodes.index(c)+index, file=link_test_n_file)
             else:
                 for c in n.get_children():
-                    print(i, self.nodes.index(c), file=link_train_file)
+                    print(i+index, self.nodes.index(c)+index, file=link_train_file)
             #for c in n.get_inter_children():
             #    G.add_edge(str(n)+" mem: "+str(id(n)), str(c)+" mem: "+str(id(c)), style="dashed")
             for c in n.get_fictitious_children():
-                print(i, self.nodes.index(c), file=link_train_file)
+                print(i+index, self.nodes.index(c)+index, file=link_train_file)
 
         link_train_file.close()
         link_test_n_file.close()
